rbd: add an option for copy-on-read
This is a new feature of rbd layered, when reading an object of child,
while it not exists, the kernel rbd client will send a new request to
parent for the origin object content, and write it to child image.

This feature is usefull when child image and its parent are geographically
isolated.

Signed-off-by: Min Chen <minchen@ubuntukylin.com>
Signed-off-by: Li Wang <liwang@ubuntukylin.com>
Signed-off-by: Yunchuan Wen <yunchuanwen@ubuntukylin.com>